ingestion/requirements.txt
kafka-python==2.0.2
websockets==12.0
python-dotenv==1.0.0
certifi
requests
playwright
playwright-stealth
